Use artifacts from bst-1.2 branch
At present the docs website uses artifacts from the master branch, which
is increasingly unhelpful for people using 1.2.x. As this is the
recommended stable branch, it makes sense for the documentation to be
for this version. This uses the latest artifact from the `bst-1.2`
branch of BuildStream, which should be fine, but it would be nicer to
get it to automatically use artifacts from the last tag on that branch,
if that's possible.

Addresses #1